refactor: complete shared public APIs
