+++ src/adapters/firestore/event.ts
@@ -9,7 +9,7 @@ import { onSnapshot, where, collection, query } from "firebase/firestore";
 import { mapCardDocument, mapDeckDocument } from "@/adapters/firestore/dto";
 import type { Card } from "@/entities/card/model/card";
 import type { Deck } from "@/entities/deck/model/deck";
-import type { RemoteChange, RemoteSubscriptionProps } from "@/shared/api/remoteSnapshot";
+import type { RemoteChange, RemoteSubscriptionProps } from "@/shared/api";
 import { getDb } from "@/shared/firebase/firestore-runtime";
 
 type RemoteEntity = { id: string; updatedAt: number; deletedAt: number | null };
