+++ src/app/providers/remote-read/RemoteReadBootstrap.integration.spec.tsx
@@ -34,7 +34,7 @@ vi.mock("@/app/providers/remote-read/remoteReadLifecycle", () => ({
 import { AuthProvider } from "@/app/providers/auth";
 import { RemoteReadBootstrap } from "@/app/providers/remote-read";
 import { createAuthRuntime } from "@/features/auth";
-import { useRemoteReadScopeUid } from "@/shared/lib/remote-read/RemoteReadScope";
+import { useRemoteReadScopeUid } from "@/shared/lib/remote-read";
 
 const ReadScopeProbe = () => <output data-testid="read-scope">{useRemoteReadScopeUid() ?? "signed-out"}</output>;
