+++ src/app/providers/remote-read/RemoteReadBootstrap.tsx
@@ -3,7 +3,7 @@ import { type PropsWithChildren, useEffect, useState } from "react";
 import { createAuthTransitionController } from "@/app/providers/auth/authTransitionController";
 import { startRemoteReads, stopRemoteReads } from "@/app/providers/remote-read/remoteReadLifecycle";
 import { useSession } from "@/entities/session";
-import { RemoteReadScopeProvider } from "@/shared/lib/remote-read/RemoteReadScope";
+import { RemoteReadScopeProvider } from "@/shared/lib/remote-read";
 
 export const RemoteReadBootstrap = ({ children }: PropsWithChildren) => {
   const session = useSession();
