+++ src/entities/deck/hooks/useDecks.ts
@@ -3,8 +3,8 @@ import { useStore } from "zustand";
 
 import type { Deck } from "@/entities/deck/model/deck";
 import { deckRemoteReadStore } from "@/entities/deck/model/remoteReadStore";
-import type { RemoteById } from "@/shared/api/remoteSnapshot";
-import { useRemoteReadScopeUid } from "@/shared/lib/remote-read/RemoteReadScope";
+import type { RemoteById } from "@/shared/api";
+import { useRemoteReadScopeUid } from "@/shared/lib/remote-read";
 
 const EMPTY_DECKS: RemoteById<Deck> = {};
