+++ src/features/card/hooks/useCardFormState.ts
@@ -9,7 +9,7 @@ import type { Card } from "@/entities/card";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 
-import type { Option } from "@/shared/ui/forms/Select";
+import type { Option } from "@/shared/ui/forms";
 import type { CardFormProps } from "@/features/card/components/CardForm";
 import { cardFormSchema, type CardFormValues } from "@/features/card/lib/cardFormSchema";
