+++ src/pages/card-form/ui/CardFormView.stories.tsx
@@ -8,7 +8,7 @@ import type { Card } from "@/entities/card";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import type { Option } from "@/shared/ui/forms/Select";
+import type { Option } from "@/shared/ui/forms";
 import type { CardFormFields } from "@/features/card";
 import * as fixture from "@/storybook/fixture";
 import { INITIAL_VIEWPORTS } from "@/storybook/storybookViewports";
