+++ src/shared/api/index.ts
@@ -1,0 +1,9 @@
+export { toRemoteById } from "./remoteSnapshot";
+export type {
+  RemoteById,
+  RemoteChange,
+  RemoteSnapshot,
+  RemoteSnapshotMetadata,
+  RemoteSubscriptionProps,
+  RemoteSyncStatus,
+} from "./remoteSnapshot";
