+++ src/shared/ui/content/index.ts
@@ -8,4 +8,5 @@ export { Section } from "./Section";
 export { Style } from "./Style";
 export { TagLabel } from "./TagLabel";
 export { TagList } from "./TagList";
+export { tagClassName } from "./tagStyles";
 export { Title } from "./Title";
